requirements.txt
-e git://github.com/ElementAI/haven.git#egg=haven
numpy~=1.19.1
torchvision~=0.7.0
pandas~=1.1.3
tqdm~=4.50.2
pillow~=7.2.0
opencv-python~=4.4.0.44
scikit-image~=0.17.2
ipython~=7.16.1
matplotlib~=3.3.2
sklearn~=0.0
scikit-learn~=0.23.2
scipy~=1.5.2
pytz~=2020.1
setuptools~=50.3.0
albumentations~=0.4.6